Use warnings not errors and add id to div:
Now when parsing units, we warn not raise if we couldn't figure out the
units so that we can keep on displaying the graph. In the case of
check_las, where we try_excepted those errors as part of the algo... we
should probably just use None but instead of raise the error.

We also add an ID to the check_las div.