Increment track index after checking for skip
We're trying to help the renderer figure out if all the children are
hidden. In this case, we can use the loop count (its manually done, not
enumerated) to see if its > -1. But this wll change depth position
mechanics since it no longer counts hidden tracks. It'll be interesting
to see if it works.