Modify ThemeStack to throw error for missing key:
There's some usability issues here.

ThemeStack would previous return None for non-existent keys.
I changed that so that if the final process sees a None, it throws a
KeyError. I did that so writing __contains__ would be better. There's a
benefit- it forces us to explicitly create defaults for each key, since
the renderers almost never check to see if the key exists. But there's a
problem, in that sometimes None is a valid value for themes. Like with
units. So now we're doing a last minute False to None version there.

It would be better to somehow differentiate between None and not
existing.

That said, if not themes['key'] is nice and probably as good as if 'key'
not in themes, although they are slightly different. I think I want to
be able to return None.

We should also really not be accessing themes['often'] because that's
sort of an operation.